feat: Add a Query challenge handler
This takes Postgres DB connection details and a query for an
action and a pattern. The action query gets given `local_part`
and `domain_part` and is expected to return a zero or one action
entries. The pattern query gets no parameters and is expected
to return zero or more pairs of a pattern and an action entry